pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-template"
version = "0.1.0"
description = "Generic Proxy-Pointer RAG FastAPI template"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "httpx>=0.27",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "uvicorn>=0.30",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "ruff>=0.5",
]
faiss = ["faiss-cpu>=1.8"]
llamaparse = ["llama-cloud-services>=0.6"]

[tool.setuptools.packages.find]
include = ["app*", "core*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "live_models: tests that call live LLM or embedding providers",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
